Split gc::ArenaIndex into its own module
Add Deref, DerefMut, AsRef, AsMut impls.
